+++ src/events.ts
@@ -1,11 +1,14 @@
-type Version = 1;
+import { Values } from "./helpers";
+
+export type Version = 1;
 
 export const EventType = {
   handshake: "handshake",
   response: "response",
 } as const;
+export type EventType = Values<typeof EventType>;
 
-type Event<Name extends keyof typeof EventType, Payload extends {}> = {
+type Event<Name extends EventType, Payload extends {}> = {
   payload: Payload;
   type: Name;
 };
@@ -35,7 +38,6 @@ export type DispatchResponseEvent = Event<
 // >;
 
 export type Events = HandshakeEvent | DispatchResponseEvent;
-export type EventType = Events["type"];
 export type PayloadOfEvent<
   TEventType extends EventType,
   TEvent extends Events = Events
