+++ src/helpers.ts
@@ -1,0 +1,1 @@
+export type Values<T> = T[keyof T];
